When selecting edit-text (for changing characteristic value), soft keyboard showed up, but the cursor didn't immediately settle inside the text-box. After setting [android:descendantFocusability="afterDescendants"] for ExpandableListView, selected child-row (descendant) started to get focused immediately. One new problem appears: edit-box of the last child of the last group doesn't get focused, and keyboard disapperas (after popping up for a while). But if the existing value of the edit box is painted first, the soft keyboard will settle down. This 'last child' problem will need some more investigation, and will be fixed later...
When clicking Android's back-button (while being connected), connection-fragment naturally will close.
But the connection remains... Added OnDestroy-method, where the connection to the remote device is now disconnected,
when back-button is clicked.

Value of the Connecting-timer in connection establishment is changed from 30 s to 15 s.

Text of the ACK-button (Confirmation) is changed according to the selected Radio-button:
  Read: "Request"
  Write: "Send"
  Notfication: button invisible

Confirmation-button moved below the edit-value box. Previously the were side to side, and when typing a lot of text,
the button moved right and finally out of the display...

Some minor layout-changes to get scanning- and connection- windows (fragments) to be closer to display-frame.